backend/requirements.txt
Django==5.1.1
djangorestframework==3.16.0
pytest==8.3.5
pytest-django==4.11.1
djoser==2.3.1
gunicorn==23.0.0
webcolors==24.11.1
Pillow==11.2.1
PyYAML==6.0.2
psycopg2-binary==2.9.10